factories and seeders

+++ database/migrations/2024_09_12_075641_create_exams_table.php
@@ -13,11 +13,13 @@ return new class extends Migration
     {
         Schema::create('exams', function (Blueprint $table) {
             $table->id();
+
             $table->unsignedBigInteger('tester_id');
             $table->foreign('tester_id')->references('id')->on('users');
+
             $table->unsignedBigInteger('student_id');
             $table->foreign('student_id')->references('id')->on('users');
-            $table->bigInteger('score');
+            $table->unsignedTinyInteger('score');
             $table->enum('result', [""]);
             $table->date('date');
             $table->unsignedBigInteger('material_id');
